Add Git server
As an initial and not exactly secure implementation of the git server stuff, i shell out to a git installation on the system. This needs to be done, because libgit2  (and as such git2-rs) does not implement the server part. This needs to be changed as soon as gitoxide has implemented this.